ATR-801: renamed namespace in tests
