+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NonPublicGraphsDacsAndExtensions/Sources/NonPublicDacExtension_Expected.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using PX.Data;
 
-namespace Acuminator.Tests.Tests.StaticAnalysis.NonPublicExtensions.Sources
+namespace Acuminator.Tests.Tests.StaticAnalysis.NonPublicGraphsDacsAndExtensions.Sources
 {
 	public sealed class SOOrderExt : PXCacheExtension<SOOrder>
 	{
