+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NonPublicGraphsDacsAndExtensions/Sources/NonPublicDac_Expected.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using PX.Data;
 
-namespace Acuminator.Tests.Tests.StaticAnalysis.NonPublicExtensions.Sources
+namespace Acuminator.Tests.Tests.StaticAnalysis.NonPublicGraphsDacsAndExtensions.Sources
 {
 	public sealed class SOOrder : IBqlTable
 	{
